requirements.txt
duckdb==0.10.3
polars==0.20.31
openpyxl==3.1.2
psycopg[binary]==3.1.19
nats-py==2.7.2
python-dateparser==1.2.0
streamlit
jinja2==3.1.4
watchdog==4.0.1
apscheduler==3.10.4
SQLAlchemy==1.4.53
requests==2.32.3
python-dotenv==1.0.1
pytest==8.1.1
portalocker
